Issue #248: Deleting account in cloud - navigation

+++ app/src/test/java/org/supla/android/data/source/DefaultChannelRepositoryTest.java
@@ -242,7 +242,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelId = 234;
-    int channelDbId = 345;
+    Long channelDbId = 345L;
 
     SuplaChannelValue suplaChannelValue = suplaChannelValue(12L);
     SuplaChannel suplaChannel =
@@ -305,7 +305,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelId = 234;
-    int channelDbId = 345;
+    Long channelDbId = 345L;
     int channelsCount = 12;
 
     SuplaChannelValue suplaChannelValue = suplaChannelValue(14L);
@@ -374,7 +374,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelId = 234;
-    int channelDbId = 345;
+    Long channelDbId = 345L;
 
     SuplaChannelValue suplaChannelValue = suplaChannelValue(12L);
     SuplaChannel suplaChannel =
@@ -551,7 +551,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelGroupId = 234;
-    int channelGroupDbId = 345;
+    Long channelGroupDbId = 345L;
 
     SuplaChannelGroup suplaChannelGroup =
         suplaChannelGroup(channelGroupId, locationId, "caption", 1, 2, 3, 4);
@@ -588,7 +588,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelGroupId = 234;
-    int channelGroupDbId = 345;
+    Long channelGroupDbId = 345L;
 
     SuplaChannelGroup suplaChannelGroup =
         suplaChannelGroup(channelGroupId, locationId, "caption1", 9, 8, 7, 6);
@@ -626,7 +626,7 @@ public class DefaultChannelRepositoryTest {
     // given
     int locationId = 123;
     int channelGroupId = 234;
-    int channelGroupDbId = 345;
+    Long channelGroupDbId = 345L;
 
     SuplaChannelGroup suplaChannelGroup =
         suplaChannelGroup(channelGroupId, locationId, "caption", 9, 8, 7, 6);
